OBPIH-983 Create API for saving stock movement items
