+++ grails-app/controllers/org/pih/warehouse/api/StockMovementApiController.groovy
@@ -12,6 +12,8 @@ package org.pih.warehouse.api
 import grails.converters.JSON
 import org.apache.commons.collections.FactoryUtils
 import org.apache.commons.collections.list.LazyList
+import org.codehaus.groovy.grails.web.json.JSONObject
+import org.pih.warehouse.core.Constants
 import org.pih.warehouse.core.Location
 import org.pih.warehouse.core.Person
 import org.pih.warehouse.inventory.InventoryItem
@@ -49,211 +51,36 @@ class StockMovementApiController {
         render ([data:stockMovement] as JSON)
 	}
 
-    def update = { StockMovement stockMovement ->
-        stockMovementService.updateStockMovement(stockMovement)
+    def update = { //StockMovement stockMovement ->
+
+        JSONObject json = request.JSON
+        log.info "json: " + json
+
+
+        def lineItems = json.remove("lineItems")
+        StockMovement stockMovement = new StockMovement()
+        bindData(stockMovement, json)
+        
+        log.info "line items: " + lineItems
+        lineItems.each { lineItem ->
+            log.info "product" + lineItem["product.id"]
+            StockMovementItem stockMovementItem = new StockMovementItem()
+            stockMovementItem.id = lineItem.id
+            stockMovementItem.deleted = lineItem.deleted ? Boolean.parseBoolean(lineItem.deleted):Boolean.FALSE
+            stockMovementItem.product = lineItem["product.id"] ? Product.load(lineItem["product.id"]) : null
+            //stockMovementItem.recipient = lineItem["recipient.id"] ? Person.load(lineItem["recipient.id"]) : null
+            stockMovementItem.quantityRequested = lineItem.quantityRequested ? new BigDecimal(lineItem.quantityRequested) : null
+            stockMovementItem.sortOrder = lineItem.sortOrder ? new Integer(lineItem.sortOrder) : null
+            stockMovement.lineItems.add(stockMovementItem)
+        }
+
+        stockMovement = stockMovementService.updateStockMovement(stockMovement)
+        render ([data:stockMovement] as JSON)
     }
 
     def delete = {
         stockMovementService.deleteStockMovement(params.id)
         render status: 204
-    }
-
-}
-
-enum StockMovementType {
-
-    INBOUND('Inbound'),
-    OUTBOUND('Outbound'),
-    OUTBOUND_STOCKLIST('Outbound with stocklist');
-
-    String name
-
-    StockMovementType(String name) { this.name = name; }
-
-    static list() {
-        [ INBOUND, OUTBOUND, OUTBOUND_STOCKLIST]
-    }
-}
-
-class StockMovement {
-
-    String id
-    String name
-    String description
-    Location origin
-    Location destination
-    Person requestedBy
-    Date dateRequested
-
-    StockMovementType stockMovementType
-
-    List<StockMovementItem> lineItems =
-            LazyList.decorate(new ArrayList(), FactoryUtils.instantiateFactory(StockMovementItem.class));
-
-    Requisition stocklist
-    Requisition requisition
-    Order order
-    Shipment shipment
-
-    static constraints = {
-        id(nullable:true)
-        name(nullable:false)
-        description(nullable:true)
-        origin(nullable:false)
-        destination(nullable:false)
-        stocklist(nullable:true)
-        requestedBy(nullable:false)
-        dateRequested(nullable:false)
-        stockMovementType(nullable:true)
-    }
-
-
-    Map getPropertyMap(String [] propertyNames) {
-        Map propertyMap = [:]
-
-        propertyNames.each { propertyName ->
-            propertyMap << ["${propertyName}": this."${propertyName}"]
-        }
-        return propertyMap;
-    }
-
-    Map toJson() {
-        return [
-                id: id,
-                name: name,
-                description: description,
-                identifier: requisition?.requestNumber,
-                origin: origin.id,
-                destination: destination.id,
-                dateRequested: dateRequested,
-                lineItems: lineItems.collect { it.toJson() }
-        ]
-    }
-
-
-    static StockMovement createFromRequisition(Requisition requisition) {
-        StockMovement stockMovement = new StockMovement(
-                id: requisition.id,
-                name: requisition.name,
-                description: requisition.name,
-                origin: requisition.origin,
-                destination: requisition.destination,
-                dateRequested: requisition.dateCreated,
-                requisition: requisition
-        )
-
-        requisition.requisitionItems.each { requisitionItem ->
-            StockMovementItem stockMovementItem = StockMovementItem.createFromRequisitionItem(requisitionItem)
-            stockMovement.lineItems.add(stockMovementItem)
-        }
-        return stockMovement
-
-    }
-
-    static StockMovement createFromShipment(Shipment shipment) {
-        StockMovement stockMovement = new StockMovement(
-                id: shipment.id,
-                name: shipment.name,
-                description: shipment.name,
-                origin: shipment.origin,
-                destination: shipment.destination,
-                dateRequested: shipment.dateCreated,
-                shipment: shipment
-        )
-
-        shipment.shipmentItems.each { shipmentItem ->
-            StockMovementItem stockMovementItem = StockMovementItem.createFromShipmentItem(shipmentItem)
-            stockMovement.lineItems.add(stockMovementItem)
-        }
-        return stockMovement
-    }
-}
-
-class StockMovementItem {
-
-    String id
-    String productCode
-    Product product
-    InventoryItem inventoryItem
-    BigDecimal quantityRequested
-    BigDecimal quantityAllowed
-    BigDecimal quantityAvailable
-    Person recipient
-
-    String palletName
-    String boxName
-
-    Integer sortOrder
-
-
-    static constraints = {
-        id(nullable:true)
-        productCode(nullable:false)
-        product(nullable:true)
-        inventoryItem(nullable:true)
-        quantityRequested(nullable:false)
-        quantityAllowed(nullable:true)
-        quantityAvailable(nullable:true)
-        recipient(nullable:true)
-        palletName(nullable:true)
-        boxName(nullable:true)
-        sortOrder(nullable:true)
-    }
-
-    Map toJson() {
-        return [
-                id: id,
-                productCode: productCode,
-                product: product,
-                palletName: palletName,
-                boxName: boxName,
-                quantityRequested: quantityRequested,
-                quantityAllowed: quantityAllowed,
-                quantityAvailable: quantityAvailable,
-                recipient: recipient,
-                sortOrder: sortOrder
-        ]
-    }
-
-    static StockMovementItem createFromShipmentItem(ShipmentItem shipmentItem) {
-
-        String palletName, boxName
-        if(shipmentItem?.container?.parentContainer) {
-            palletName = shipmentItem?.container?.parentContainer?.name
-            boxName = shipmentItem?.container?.name
-        } else if (shipmentItem.container) {
-            palletName = shipmentItem?.container?.name
-        }
-
-        return new StockMovementItem(id: shipmentItem?.id,
-                productCode: shipmentItem?.product?.productCode,
-                product: shipmentItem?.inventoryItem?.product,
-                inventoryItem: shipmentItem?.inventoryItem,
-                quantityRequested: shipmentItem?.quantity,
-                quantityAllowed: null,
-                quantityAvailable: null,
-                palletName:palletName,
-                boxName:boxName,
-                recipient: shipmentItem.recipient,
-                sortOrder: null
-
-        )
-    }
-
-    static StockMovementItem createFromRequisitionItem(RequisitionItem requisitionItem) {
-        return new StockMovementItem(id: requisitionItem.id,
-                productCode: requisitionItem?.product?.productCode,
-                product: requisitionItem?.product,
-                inventoryItem: requisitionItem?.inventoryItem,
-                quantityRequested: requisitionItem.quantity,
-                quantityAllowed: null,
-                quantityAvailable: null,
-                palletName:null,
-                boxName:null,
-                recipient: requisitionItem.recipient,
-                sortOrder: requisitionItem.orderIndex
-
-        )
     }
 
 }
