+++ grails-app/services/org/pih/warehouse/inventory/StockMovementService.groovy
@@ -13,6 +13,7 @@ import grails.validation.ValidationException
 import org.apache.commons.lang.NotImplementedException
 import org.hibernate.ObjectNotFoundException
 import org.pih.warehouse.api.StockMovement
+import org.pih.warehouse.api.StockMovementItem
 import org.pih.warehouse.requisition.Requisition
 import org.pih.warehouse.requisition.RequisitionItem
 import org.pih.warehouse.requisition.RequisitionStatus
@@ -37,8 +38,8 @@ class StockMovementService {
         // perspective of the requisition, while the origin/destination of the stock movement related to the stock
         // being transferred.
         requisition.status = RequisitionStatus.CREATED
-        requisition.destination = stockMovement.origin
-        requisition.origin = stockMovement.destination
+        requisition.destination = stockMovement.destination
+        requisition.origin = stockMovement.origin
         requisition.name = stockMovement.name
         requisition.description = stockMovement.description
         requisition.requestedBy = stockMovement.requestedBy
@@ -63,7 +64,73 @@ class StockMovementService {
     }
 
     def updateStockMovement(StockMovement stockMovement) {
-        throw new NotImplementedException("Update stock movement has not been implemented")
+        log.info "Update stock movement " + stockMovement.lineItems
+
+        if (!stockMovement.validate()) {
+            throw new ValidationException("Invalid stock movement", stockMovement.errors)
+        }
+
+        Requisition requisition = Requisition.get(stockMovement.id)
+        if (!requisition) {
+            requisition = new Requisition()
+        }
+
+        if (stockMovement.destination) requisition.destination = stockMovement.destination
+        if (stockMovement.origin) requisition.origin = stockMovement.origin
+        if (stockMovement.name) requisition.name = stockMovement.name
+        if (stockMovement.description) requisition.description = stockMovement.description
+        if (stockMovement.requestedBy) requisition.requestedBy = stockMovement.requestedBy
+        if (stockMovement.dateRequested) requisition.dateRequested = stockMovement.dateRequested
+        //if (stockMovement.identifier) requisition.requestNumber = stockMovement.identifier
+
+        if (stockMovement.lineItems) {
+            stockMovement.lineItems.each { StockMovementItem stockMovementItem ->
+                RequisitionItem requisitionItem
+                // Try to find a matching stock movement item
+                if (stockMovementItem.id) {
+                    requisitionItem = requisition.requisitionItems.find { it.id == stockMovementItem.id }
+                }
+
+
+                // If requisition item is found, we update it
+                if (requisitionItem) {
+                    log.info "Item found " + requisitionItem.id
+
+                    if (stockMovementItem.deleted) {
+                        log.info "Item deleted " + requisitionItem.id
+                        requisition.removeFromRequisitionItems(requisitionItem)
+                        requisitionItem.delete()
+                    }
+                    else {
+                        log.info "Item updated " + requisitionItem.id
+                        if (stockMovementItem.product) requisitionItem.product = stockMovementItem.product
+                        if (stockMovementItem.inventoryItem) requisitionItem.product = stockMovementItem.inventoryItem
+                        if (stockMovementItem.quantityRequested) requisitionItem.quantity = stockMovementItem.quantityRequested
+                        //if (stockMovementItem.recipient) requisitionItem.recipient = stockMovementItem.recipient
+                        if (stockMovementItem.sortOrder) requisitionItem.orderIndex = stockMovementItem.sortOrder
+                    }
+                    requisitionItem.save()
+                }
+                // Otherwise we create a new one
+                else {
+                    log.info "Item not found"
+                    requisitionItem = new RequisitionItem()
+                    requisitionItem.product = stockMovementItem.product
+                    requisitionItem.inventoryItem = stockMovementItem.inventoryItem
+                    requisitionItem.quantity = stockMovementItem.quantityRequested
+                    //requisitionItem.recipient = stockMovementItem.recipient
+                    requisitionItem.orderIndex = stockMovementItem.sortOrder
+                    requisition.addToRequisitionItems(requisitionItem)
+                }
+            }
+        }
+
+
+        if (requisition.hasErrors() || !requisition.save(flush:true)) {
+            throw new ValidationException("Invalid requisition", requisition.errors)
+        }
+
+        return StockMovement.createFromRequisition(requisition)
     }
 
     def deleteStockMovement(String id) {
